EDU-335 Reduce context injection to repository-browser only
