+++ src/components/page-name-context.js
@@ -8,15 +8,7 @@ export function usePageName() {
   return useContext(pageNameContext);
 }
 
-export function withPageName(Component) {
-  return function RequestInjector(props) {
-    const pageName = useContext(pageNameContext);
-    return <Component {...props} pageName={pageName} />;
-  };
-}
-
 export default {
   PageNameProvider,
-  usePageName,
-  withPageName
+  usePageName
 };
