+++ src/components/request-context.js
@@ -8,15 +8,7 @@ export function useRequest() {
   return useContext(requestContext);
 }
 
-export function withRequest(Component) {
-  return function RequestInjector(props) {
-    const request = useContext(requestContext);
-    return <Component {...props} request={request} />;
-  };
-}
-
 export default {
   RequestProvider,
-  useRequest,
-  withRequest
+  useRequest
 };
