+++ src/components/settings-context.js
@@ -8,15 +8,7 @@ export function useSettings() {
   return useContext(settingsContext);
 }
 
-export function withSettings(Component) {
-  return function RequestInjector(props) {
-    const settings = useContext(settingsContext);
-    return <Component {...props} settings={settings} />;
-  };
-}
-
 export default {
   SettingsProvider,
-  useSettings,
-  withSettings
+  useSettings
 };
